First day of work. Heart Defense prototype
Have pathways to a heart model with viral pawns that move towards the
heart upon spawn (spawning from recurring timers on factory actors).
However, the central pathway results in interrupted movement somehow for
the pawns. Also, beginning to think that there are some serious design
issues with the fundamental concept of this game design. I simply don't
have fun thinking about what we'd be doing in-game. And if I can't look
forward to it, that's a problem.